Rename VoictentConfiguration file to Voque
