+++ packages/voictents-and-estinants-engine/src/core/engine/voque.ts
